+++ large-project/src/App.js
@@ -4,7 +4,6 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import LoginPage from './pages/LoginPage';
 import SignupPage from './pages/SignupPage';
 import ForgetPage from './pages/ForgetPage';
-import CardPage from './pages/CardPage';
 import MapsPage from './pages/MapsPage';
 function App() {
   return (
@@ -13,7 +12,6 @@ function App() {
       <Route path="/" index element={<LoginPage />} />
       <Route path="/Signup" index element={<SignupPage />} />
       <Route path="/ForgetPage" index element={<ForgetPage />} />
-      <Route path="/cards" index element={<CardPage />} />
       <Route path="/map" index element={<MapsPage />} />
     </Routes>
   </BrowserRouter>
